packaging_automation/requirements.txt

adal==1.2.2
appdirs==1.4.4
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
Deprecated==1.2.12
devscripts===2.20.2ubuntu2
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
docker==5.0.0
entrypoints==0.3
filelock==3.0.12
gitdb==4.0.7
GitPython==3.1.14
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
iniconfig==1.1.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
packaging==20.9
parameters-validation==1.2.0
pathlib2==2.3.5
pbr==5.5.1
pexpect==4.6.0
pipenv==2020.11.15
pluggy==0.13.1
py==1.10.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.20.0
pycurl==7.43.0.6
PyGithub==1.54.1
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.15.5
pyserial==3.4
pytest==6.2.3
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-gnupg==0.4.7
python-magic==0.4.17
python-string-utils==1.0.0
pyxdg==0.26
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
smmap==4.0.0
sos==4.1
ssh-import-id==5.10
systemd-python==234
testresources==2.0.1
toml==0.10.2
Twisted==18.9.0
ubuntu-advantage-tools==20.3
ufw==0.36
unidiff==0.5.5
urllib3==1.25.8
virtualenv==20.0.23
virtualenv-clone==0.5.4
wadllib==1.3.3
websocket-client==0.59.0
wrapt==1.12.1
zipp==1.0.0
zope.interface==4.7.1